UISScreen.c: fn_8016A2D4 95.83 -> 96.03 (fn_8016C270 returns s32)

+++ src/UISScreen.c
@@ -150,7 +150,7 @@ u8* fn_8016C5C4(UISNode* pNode, u32 uEvent) {
 // Pushes a call frame on pStack (the saved word, the extra word, both argument lists, the node's
 // info and a 0) and runs pScript on it. The frame stays on the stack only when the script
 // returns 3 (it paused).
-s8 fn_8016C270(UIStudio* pStudio, UISScreen* pScreen, UISNodeInfo* pInfo, UISWordStack* pStack, u8* pScript,
+s32 fn_8016C270(UIStudio* pStudio, UISScreen* pScreen, UISNodeInfo* pInfo, UISWordStack* pStack, u8* pScript,
                s32 nArgs, const s32* pArgs, u32 nArgs2, const s32* pArgs2, u8 bExtra, s32 nExtra,
                s32* pnSaved) {
     s32* pFrame;
